requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.1
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pydantic>=2.6.0
python-dotenv>=1.0.0
langchain>=0.0.292
groq>=0.4.0
langchain-groq>=0.0.1
langchain-core>=0.1.0
streamlit>=1.28.0
requests>=2.31.0
